Use isolated Lua script for brick-and-mortar
Since there will be a number of CLI parameters for the brick-and-mortar
Lua benchmarking script that won't be relevant to the employee-directory
benchmarks, make the bench.lua into brick-and-mortar.lua and remove the
application CLI option.

Also corrects some problems in the schema files and adds creation of the
schema and cleanup of the schema to the Lua script.

Issue #4