Rework brick-and-mortar prepare() function
After encountering a number of issues with using a SQL script with
multiple statements for loading schema, decided to rework things to have
an in-script table of SQL statements. This got rid of the issues where I
was getting a 2014: Commands out of sync error from MySQL using a
single connection result.

Issue #4