Use Lua UUID generation
Instead of using the MySQL UUID() function to generate a UUID key for
things like orders, we are going to need to generate the UUID ahead of
time and then pass the UUID value to the prepared statements for
order_details (since we'll need the UUID of the order as a foreign key
relationship to the order_details table).

Unfortunately, this means using a Lua UUID generator for now at least.
Consider adding UUID generation utility to sysbench's C code.

Issue #4