Remove employee directory application
Adds an order_counts_by_status scenario against the brick-and-mortar
application after adding an index on orders.status. This rounded out the
archetypal queries I wanted to examine and therefore I have removed the
employee directory application and am going to just go with the
brick-and-mortar store application to illustrate the impact of column
type choice of primary key.

Issue #4