Correct PostgreSQL schema design C indexes
After analyzing the numbers for the PostgreSQL schema design "C"
scenario results, I saw an anomaly that seemed odd. The benchmark
numbers for the schema design "C" were much lower than design "B" or
"A", especially for the lookup_orders_by_customer scenario.

I setup DBs for both the "C" and "B" schema designs and ran EXPLAIN
SELECT on the particular query issued in the lookup_orders_by_customer
scenario and noted that a sequential scan of the customers table was
being done when an index on the customers.uuid column should have been
used. Turns out I had forgotten to manually create the indexes on the
uuid columns for PostgreSQL and schema design "C". So, I added those
indexes into the prepare stage and re-ran the benchmarks for the "C"
schema design.

This patch updates the images and results after that run.

Issue #4